requirements.txt
torch>=1.9.0
transformers>=4.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
plotly>=5.0.0
flask>=2.0.0
pyyaml>=5.4.0
tqdm>=4.60.0
PyPDF2>=2.0.0
python-docx>=0.8.11
sentencepiece>=0.1.96
protobuf>=3.19.0